Find builtin module automatically
