+++ trytune/services/moduels/__init__.py
@@ -1,45 +1,29 @@
-from abc import ABC, abstractmethod
+import inspect
 from typing import Any, Dict
 
-import numpy as np
-
-from trytune.schemas.common import DataSchema
-
-
-class BuiltinModule(ABC):
-    """
-    BuiltinModule is a base class for all builtin modules. Builtin modules are executed in API server.
-    It generally run on CPU and is used for post-processing, e.g., NMS, etc. But it can also run on GPU if API server has GPU.
-
-    Builtin module has three methods:
-    - initialize: Initialize builtin module. This method is called when API server starts.
-    - execute: Execute builtin module. This method is called when API server receives a request.
-    - metadata: Return metadata of builtin module. This method is called when API server starts.
-    Note that all methods are async.
-
-    All builtin modules are must be implemented in trytune/services/modules/builtin.py.
-    When API server starts, all builtin modules will be automatically collected and registered into moudles.
-    Note that builtin modules cannot be dynamically loaded.
-    """
-
-    # E.g., {"name": "nms", "args": {"threshold": 0.9}}
-    @abstractmethod
-    async def initialize(self, args: Dict[str, Any]) -> None:
-        raise NotImplementedError
-
-    @abstractmethod
-    async def execute(self, requests: Any) -> Dict[str, np.ndarray]:
-        raise NotImplementedError
-
-    @abstractmethod
-    def metadata(self) -> Dict[str, Any]:
-        raise NotImplementedError
+from trytune.services.moduels import builtin, common
 
 
 # Class to store module metadatas and links to triton servers.
 class Modules:
     def __init__(self) -> None:
+        # All modules installed in the system.
         self.modules: Dict[str, Dict[str, Any]] = {}
+        # All builtin modules available in the system. But, not yet installed.
+        self.available_builtins: Dict[str, Any] = {}
+
+        for name, obj in inspect.getmembers(builtin):
+            if (
+                inspect.isclass(obj)
+                and issubclass(obj, common.BuiltinModule)
+                and name != "BuiltinModule"
+            ):
+                metadata = obj().metadata()
+                metadata["name"] = name
+                metadata["is_builtin"] = True
+                self.available_builtins[name] = metadata
+                # FIXME: convert to debug log
+                print(f"Found builtin module: {name}: {metadata}")
 
     def set(self, module: str, metadata: Dict[str, Any]) -> None:
         assert module not in self.modules
